Move DriverPlugin interface back to its prior location
